Feature/218 npc target selection based on threat

+++ MysteryMud.Domain/Action/Attack/Resolvers/IAggroResolver.cs
@@ -6,6 +6,6 @@ namespace MysteryMud.Domain.Action.Attack.Resolvers;
 
 public interface IAggroResolver
 {
-    void ResolveFromDamage(GameState state, Entity target, Entity source, int damageAmount, DamageKind damageKind);
-    void ResolveFromHeal(GameState state, Entity target, Entity source, int healAmount);
+    void ResolveFromDamage(GameState state, Entity target, Entity source, decimal damageAmount, DamageKind damageKind);
+    void ResolveFromHeal(GameState state, Entity target, Entity source, decimal healAmount);
 }
